Add abstract class IamRequestDTO
