+++ iam-login-service/src/main/java/it/infn/mw/iam/api/requests/model/GroupRequestDto.java
@@ -29,31 +29,17 @@ import it.infn.mw.iam.api.validators.IamGroupRequestNotes;
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 @GroupRequest
-public class GroupRequestDto {
-
-  private String uuid;
-
-  private String username;
-  
-  private String userUuid;
-  
-  private String userFullName;
+public class GroupRequestDto extends IamRequestDTO {
 
   @NotEmpty
   @IamGroupName(message = "Invalid membership request: group does not exist")
   private String groupName;
-  
-  private String groupUuid;
 
-  private String status;
+  private String groupUuid;
 
   @NotEmpty
   @IamGroupRequestNotes(message = "Invalid membership request: notes cannot be empty")
   private String notes;
-
-  private String motivation;
-  private Date creationTime;
-  private Date lastUpdateTime;
 
   public GroupRequestDto() {
     // empty constructor
@@ -61,42 +47,18 @@ public class GroupRequestDto {
 
   @JsonCreator
   public GroupRequestDto(@JsonProperty("uuid") String uuid,
-      @JsonProperty("userUuid") String userUuid,
-      @JsonProperty("userFullName") String userFullName,
+      @JsonProperty("userUuid") String userUuid, @JsonProperty("userFullName") String userFullName,
       @JsonProperty("username") String username, @JsonProperty("group_name") String groupName,
-      @JsonProperty("group_uuid") String groupUuid,
-      @JsonProperty("status") String status, @JsonProperty("notes") String notes,
-      @JsonProperty("motivation") String motivation,
+      @JsonProperty("group_uuid") String groupUuid, @JsonProperty("status") String status,
+      @JsonProperty("notes") String notes, @JsonProperty("motivation") String motivation,
       @JsonProperty("creation_time") Date creationTime,
       @JsonProperty("last_update_time") Date lastUpdateTime) {
-    
-    this.uuid = uuid;
-    this.userUuid = userUuid;
-    this.username = username;
-    this.userFullName = userFullName;
+
+    super(uuid, userUuid, userFullName, username, notes, status, motivation, creationTime,
+        lastUpdateTime);
     this.groupName = groupName;
     this.groupUuid = groupUuid;
-    this.status = status;
     this.notes = notes;
-    this.motivation = motivation;
-    this.creationTime = creationTime;
-    this.lastUpdateTime = lastUpdateTime;
-  }
-
-  public String getUuid() {
-    return uuid;
-  }
-
-  public void setUuid(String uuid) {
-    this.uuid = uuid;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
   }
 
   public String getGroupName() {
@@ -107,66 +69,20 @@ public class GroupRequestDto {
     this.groupName = groupName;
   }
 
-  public String getStatus() {
-    return status;
-  }
-
-  public void setStatus(String status) {
-    this.status = status;
-  }
-
+  @Override
   public String getNotes() {
     return notes;
   }
 
+  @Override
   public void setNotes(String notes) {
     this.notes = notes;
   }
 
-  public String getMotivation() {
-    return motivation;
-  }
-
-  public void setMotivation(String motivation) {
-    this.motivation = motivation;
-  }
-
-  public Date getCreationTime() {
-    return creationTime;
-  }
-
-  public void setCreationTime(Date creationTime) {
-    this.creationTime = creationTime;
-  }
-
-  public Date getLastUpdateTime() {
-    return lastUpdateTime;
-  }
-
-  public void setLastUpdateTime(Date lastUpdateTime) {
-    this.lastUpdateTime = lastUpdateTime;
-  }
-  
-  public String getUserUuid() {
-    return userUuid;
-  }
-  
-  public void setUserUuid(String userUuid) {
-    this.userUuid = userUuid;
-  }
-  
-  public String getUserFullName() {
-    return userFullName;
-  }
-  
-  public void setUserFullName(String userFullName) {
-    this.userFullName = userFullName;
-  }
-  
   public String getGroupUuid() {
     return groupUuid;
   }
-  
+
   public void setGroupUuid(String groupUuid) {
     this.groupUuid = groupUuid;
   }
